Add a python script to build a package in COPR
This script attempts to start parallel builds for as many dependencies
as it can. It's kind of fragile; there's no handling of errors from
threads or anything, and there's no clean way to recover from errors.

My copr project name is hard-coded in, so change that if you want to use
this script.