Fix the handling of scoped packages.
When fetching data from the npm registry, I'm not sure what is supposed
to happend to get the single-version data for a scoped package, so just
do one fetch for the package data and use data.versions.

On the template side, be sure to remove '@' from package names, and
split the scope out into a separate variable to make creating and owning
the scope directory more straightforward.