Add a --add-dep option
There are some packages out there with incorrect or incomplete
dependency data in package.json, and due to the kind of insane way that
npm installs things by default, no one notices. On top of that, none of
them are ever going to be fixed because they're from two major versions
ago but you still need the busted old thing as part of another
dependency tree.

To work around this, add a CLI option to manually patch in new
dependencies when generating a package.