Fix a race in creating new package events
When looking for an event correpsonding to a package build, if one is
not found, create the new event while the dictionary lock is held.
Otherwise there's a gap between looking for the event and creating the
new one at the start of processSRPM, and another thread needing the same
dep could create a duplicate build in the meantime.