Add an option to disable %check
Sometimes the %check target isn't going to work, and that's fine, and
there's no way to detect it. For instance, ast-types-flow contains an
es6 export statement that cannot be read by nodejs, and there is no
particular indication (I think, at least) in package.json that this is
the case, and yet this module is used as a dependency and somehow things
work. Add an option to just give up on %check for such modules.