Switch a bunch of callbacks to promises + async/await
Ditch a bunch of the callback-style code because this aggresively
asynchronous style of Javascript is annoying and gross and incredibly
pointless since everything is still sequential, it's just that the next
step gets to to be indented an extra step or two.

The remaining callback is the one used with read-package-json, since
that would be inconvenient to replace.

To get more promisey interfaces, make some changes in the dependencies:

  * switch from npm-registry-client to npm-registry-fetch. -fetch is the
    new cool one apparently.
  * switch from tar-fs to tar
  * switch from request to node-fetch

For tmp, just use the synchronous interface, and for read-package-json
add a Promise wrapper.

One downside to using Promises is holy shit did Javascript screw up
error handling on these. To get semi-sensible behavior, replace the
top-level exception catcher that no longer works with a
'unhandledRejection' event handler.